echo-api-server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
delta-spark==3.3.0
deltalake==0.16.4
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
pandas==2.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pyarrow-hotfix==0.6
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zipp==3.21.0
